[nuxt.config] Use prefix_except_default strategy
`prefix_and_default` strategy does not seem to be functional.
We had the issue reported in [1]: infinite loop on routing to a page in
the default language.

Even when `prefix_and_default` was on, the prefixed `ja` routes were not
generated, for some reason.
So we have no less in functionaly in using the `prefix_except_default`
strategy.

[1] https://github.com/nuxt-community/i18n-module/issues/1292